pyproject.toml
[project]
name = "dbos-openai"
version = "0.1.0"
description = "DBOS integration for the OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dbos>=2.10.0",
    "openai-agents>=0.7.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
filter_files = true
atomic = true

[tool.mypy]
strict = true

[tool.pytest.ini_options]
addopts = "-s"
log_cli_format = "%(asctime)s [%(levelname)8s] (%(name)s:%(filename)s:%(lineno)s) %(message)s"
log_cli_level = "INFO"
log_cli = true